+++ obs/api/app/controllers/api/storage.py
@@ -7,7 +7,7 @@ from obs.libs import gmt
 from obs.libs import auth
 from obs.libs import utils
 from requests_aws4auth import AWS4Auth
-from app.helpers.rest import response
+from obs.api.app.helpers.rest import response
 from werkzeug.utils import secure_filename
 from flask import request, send_file
 from flask_restful import Resource, reqparse
@@ -114,7 +114,7 @@ class bucket_api(Resource):
             if responses.text:
                 error = xmltodict.parse(responses.text)
                 return response(400, error["Error"]["Message"])
-            return response(201)
+            return response(201, f"Bucket {bucket_name} created successfully.")
         except Exception as exc:
             return response(500, str(exc))
 
@@ -126,10 +126,10 @@ class bucket_api(Resource):
         secret_key = args["secret_key"].replace(" ", "+")
 
         try:
-            bucket.remove_bucket(
+            result = bucket.remove_bucket(
                 get_resources(args["access_key"], secret_key), bucket_name
             )
-            return response(204)
+            return response(200, f"Bucket {bucket_name} deleted successfully.", result)
         except Exception as exc:
             return response(500, str(exc))
 
@@ -164,12 +164,14 @@ class object_api(Resource):
         secret_key = args["secret_key"].replace(" ", "+")
 
         try:
-            bucket.remove_object(
+            result = bucket.remove_object(
                 get_resources(args["access_key"], secret_key),
                 bucket_name,
                 args["object_name"],
             )
-            return response(204)
+            return response(
+                200, f"Object {args['object_name']} deleted successfully.", result
+            )
         except Exception as exc:
             return response(500, str(exc))
 
@@ -192,7 +194,7 @@ class move_object(Resource):
                 args["move_to"],
                 None,
             )
-            return response(204)
+            return response(201, f"Object {args['object_name']} moved successfully.")
         except Exception as exc:
             return response(500, str(exc))
 
@@ -215,7 +217,7 @@ class copy_object(Resource):
                 args["copy_to"],
                 None,
             )
-            return response(204)
+            return response(201, f"Object {args['object_name']} copied successfully.")
         except Exception as exc:
             return response(500, str(exc))
 
@@ -258,7 +260,7 @@ class upload_object(Resource):
         object_name = args["object_name"] if args["object_name"] else filename
 
         try:
-            bucket.upload_object(
+            result = bucket.upload_object(
                 resource=get_resources(args["access_key"], secret_key),
                 bucket_name=bucket_name,
                 local_path=filename,
@@ -274,7 +276,7 @@ class upload_object(Resource):
                     acl_type="object",
                     acl=args["acl"],
                 )
-            return response(201)
+            return response(201, f"Object {object_name} uploaded successfully.", result)
         except Exception as exc:
             return response(500, str(exc))
 
@@ -330,16 +332,17 @@ class acl(Resource):
 
         acl_type = "object" if args["object_name"] else "bucket"
         acl = args["acl"] if args["acl"] else "private"
+        name = args["object_name"] if args["object_name"] else args["bucket_name"]
 
         try:
-            bucket.set_acl(
+            result = bucket.set_acl(
                 resource=get_resources(args["access_key"], secret_key),
                 bucket_name=args["bucket_name"],
                 object_name=args["object_name"],
                 acl_type=acl_type,
                 acl=acl,
             )
-            return response(204)
+            return response(200, f"Added {acl} access to {acl_type} {name}.", result)
         except Exception as exc:
             return response(500, str(exc))
 
@@ -375,12 +378,14 @@ class mkdir(Resource):
         secret_key = args["secret_key"].replace(" ", "+")
 
         try:
-            bucket.mkdir(
+            result = bucket.mkdir(
                 get_resources(args["access_key"], secret_key),
                 bucket_name,
                 args["directory"],
             )
-            return response(201)
+            return response(
+                201, f"Directory {args['directory']} added successfully.", result
+            )
         except Exception as exc:
             return response(500, str(exc))