+++ CHANGELOG.rst
@@ -1,6 +1,40 @@
 Unreleased
 ==========
 
+0.2.2 (2020-02-12)
+==================
+- Added API to neo-obs
+    - Added list bucket in storage function
+    - Added list object in storage function
+    - Added create bucket in storage function
+    - Added delete bucket in storage function
+    - Added bucket info in storage function
+    - Added delete object in storage function
+    - Added object info in storage function
+    - Added upload object in storage function
+    - Added download object in storage function
+    - Added copy object in storage function
+    - Added move object in storage function
+    - Added object usage in storage function
+    - Added bucket usage in storage function
+    - Added set object ACL in storage function
+    - Added set bucket ACL in storage function
+    - Added get object url in storage function
+    - Added make directory in storage function
+    - Added list gmt_policy in storage function
+    - Added create user in admin function 
+    - Added delete user in admin function 
+    - Added suspend & unsuspend user in admin function 
+    - Added list user in admin function 
+    - Added user info in admin function 
+    - Added set qos user limit in admin function 
+    - Added delete qos user limit in admin function 
+    - Added qos user info in admin function
+    - Added create new credibility for user in admin function
+    - Added change status credibility for user in admin function
+    - Added delete credibility for user in admin function
+    - Added list credibility user in admin function
+
 0.1.1 (2020-01-13)
 ==================